requirements.txt
torch>=2.0.0
transformers>=4.35.0
trl>=0.7.1
numpy>=1.22.0
scipy>=1.10.0
pyyaml>=6.0
matplotlib>=3.5.0